there are too many conditional statements (like isEditing ? ) in the EditBlogForm component, which makes it not readable.
so i separate EditBlogForm into two component, one for display and another for edit.

there should be a hover menu with a "three-dot" icon on the top right of the display component,  when hover and click on it there are options of edit and delete shown.

for the edit component, there is cancel and save button on the bottom right of the component.

(and do not display the title or edit the title, the title entry is abandoned)