when click on the user information in the sidebar, there will be a dropdown menu(using Dropdown in antd), the menu should contain the following labels:  账号偏好(下有小字： 语言/偏好/存储/导出)， then a divider, then 会员中心， 记录统计， 消息统计，帮助教程， then another divider, 退出（logout）.
To the right of the user information, there should be a button to collapse sidebar, hover on it will show a tooltip "隐藏侧边栏" and click on the button will hide the sidebar. After hidding the sidebar, a burger button should be shown to the left of the Breadcrumb, clicking on the burger button will show the sidebar, while hide the burger button itself.